hnswsq: sq8/sq16 fixed range is [0, 255] (byte domain)
The user chose the byte-data range: sq8 stores round(clamp(x, 0, 255)) as an
unsigned byte (scale 1.0) and sq16 stores round(clamp(x, 0, 255) * 128) at
scale 2^-7.  BIGANN/SIFT-style uint8 vectors now encode losslessly, so the
integer pairwise distance SUM (qhat - code)^2 is the decoded-domain L2 of the
quantized query exactly on the stored side, and the layouts land plain-class
recall on the 1M benchmark with no calibration.

Kernels return to unsigned byte codes (cvtepu8, like the calibrated f8's code
space); tests switch to byte-domain data (recall/incremental/vacuum/clamp
scaffolds affine-map the clustered generator into [1, 255]).

+++ pgvectorscale/src/access_method/distance/mod.rs
@@ -196,10 +196,11 @@ pub fn distance_l2_sq8_pairwise_scalar(qhat: &[i16], code: &[u8], w: &[f32]) -> 
 }
 
 /// Fixed-range (training-free) `sq8` pairwise: `SUM (qhat - code)^2` over the
-/// int8 codes, pure i32 integer accumulation.  Safe up to the 16000-dim
-/// limit: `16000 * 254^2 ≈ 1.03e9 < i32::MAX`.  The quantization step is one
-/// global constant, so this is the decoded-domain L2 of the quantized query
-/// times `127^2` — the stored-code side is exact.
+/// unsigned byte codes, pure i32 integer accumulation.  Safe up to the
+/// 16000-dim limit: `16000 * 255^2 ≈ 1.04e9 < i32::MAX`.  The quantization
+/// step is one global constant (scale 1.0), so for byte-valued vectors this
+/// is the decoded-domain L2 of the quantized query — the stored side is
+/// exact.
 #[inline]
 pub fn distance_l2_sq8_fixed_pairwise(qhat: &[i16], code: &[u8]) -> f32 {
     #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
@@ -215,17 +216,16 @@ pub fn distance_l2_sq8_fixed_pairwise_scalar(qhat: &[i16], code: &[u8]) -> f32 {
     debug_assert_eq!(qhat.len(), code.len());
     let mut acc = 0i32;
     for i in 0..qhat.len() {
-        // Fixed sq8 codes are signed two's complement bytes.
-        let d = qhat[i] as i32 - code[i] as i8 as i32;
+        let d = qhat[i] as i32 - code[i] as i32;
         acc += d * d;
     }
     acc as f32
 }
 
 /// Fixed-range (training-free) `sq16` pairwise: `SUM (qhat - code)^2` over
-/// the int16 codes, i64 accumulation (per-dimension differences up to 65534
+/// the u16 codes, i64 accumulation (per-dimension differences up to 65280
 /// square past i32).  Equals the decoded-domain L2 of the quantized query
-/// times `32767^2`.
+/// times `128^2`.
 #[inline]
 pub fn distance_l2_sq16_fixed_pairwise(qhat: &[i16], code: &[u8]) -> f32 {
     #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
